+++ NEWS.md
@@ -1,5 +1,8 @@
 # workflows (development version)
 
+* Improved error message in `workflow_variables()` if either `outcomes` or
+  `predictors` are missing (#144).
+
 * Removed ellipsis dependency in favor of equivalent functions in rlang.
 
 * New `extract_parameter_set_dials()` and `extract_parameter_dials()` methods